requirements.txt
# SAP Audit Tool Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
pandas>=1.3.0
numpy>=1.20.0
openpyxl>=3.0.7
xlrd>=2.0.1
pytz>=2021.1

# Configuration support
python-dotenv>=0.19.0
pyyaml>=6.0

# Data analysis and visualization
matplotlib>=3.5.0

# Optional dependencies
# cx_Oracle>=8.0.0  # For direct Oracle/SAP database connections
# requests>=2.25.0  # For API integrations
# boto3>=1.17.0     # For AWS S3 storage integration
